Displaying Score and Text
Added Two Text Objects and added functionality to the PlayerController script to dynamically either update the text objects or display them. The ScoreText counts the score and the WinText will display a win message when you collect all the PickUps. Also added a GameObject to track the parent of the PickUps and count the number of children so I don't have to manually change the number till the player wins if I add a new PickUp to the map.