Refactored Pick Up Sound and Added WIn Sound
Created script to manage the sound that is played on different actions for the player. It lets the player ball hold the audio and play oneshot sounds when a condition happens. For example when the play collects a pickup, it then uses the sound manager script to play the pick up sound, and when the player wins it plays the win sound.